Move errors to mapper
