+++ src/ws/service/mapper/mod.rs
@@ -1,2 +1,3 @@
+pub mod error;
 pub mod events;
 pub mod payload_to_response;
